Skip highlighting symbols not actually in file
This resolves #675, where cquery would emit multiple `Bad index_line`
errors because it was trying to use `-1` as an index. The cause was
symbols found in a translation unit that weren't visibly present in
the file (e.g. expanded GoogleTest templated macros), and so had a
`start_line` of `-1`. The fix is to simply not publish a semantic
highlight for a symbol that cannot be highlighted.