Hash argument lists in IndexFiles
On large projects where thousands of files each have thousands of arguments,
the args array of IndexFiles ends up causing cquery to run out of memory.

Currently the array is only used to check whether it can reload indices from cache,
so let it just store a hash and compare that instead of the full array.